services/requirements.txt
# ===== CORE FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===== DATABASE & ORM =====
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
sqlalchemy-utils

# ===== AUTHENTICATION & SECURITY =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT

# ===== DATA PROCESSING =====
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# ===== MACHINE LEARNING =====
scikit-learn==1.3.2
prophet==1.1.4
statsmodels==0.14.0

# ===== VISUALIZATION =====
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# ===== HTTP & API CLIENTS =====
httpx==0.25.2
aiohttp==3.9.1
requests

# ===== PAYMENTS =====
stripe>=5.0.0

# ===== AI & LLM =====
langchain
langchain-groq
langchain-community
langchain-experimental
groq
openai
tiktoken
google-search-results
langgraph

# ===== REPORT GENERATION =====
jinja2==3.1.2
weasyprint==60.2
openpyxl==3.1.2
reportlab

# ===== UTILITIES =====
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
schedule
psutil

# ===== MONITORING & LOGGING =====
structlog==23.2.0
prometheus-client==0.19.0

# ===== TESTING (Development Only) =====
pytest==7.4.3
pytest-asyncio==0.21.1

# ===== DEVELOPMENT TOOLS (Development Only) =====
pip-tools==7.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0